Update and improve config options
PREFIXSHORCUT now defaults to PREFIX, so it is not needed anymore.

Setting FFT_LIBDIR makes $FFT_LIBDIR/include to be added to INCLUDEPATH
and -L$FFT_LIBDIR/lib to be added to libs. As it was set to the same as
PREFIX, it ended masking an issue that $PREFIX/include and
-L$PREFIX/lib were not included in the respective variables (i.e. not
setting FFT_LIBDIR also made libjack.so to not be found). So, instead of
setting FFT_LIBDIR, now we explicit configure $PREFIX/include and
-L$PREFIX/lib.